Modify player description
